# backend/internal/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Nanyi CRM 管理后台接口文档。统一前缀 /api/v1，响应结构 { code, message, data }，分页结构 { records, total, pageNum, pageSize }。异常码段：1xxx 参数 / 2xxx 认证授权 / 3xxx 业务 / 5xxx 系统。",
        "title": "Nanyi CRM API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/system/api": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需要权限：system:api:list",
                "tags": [
                    "系统管理-接口"
                ],
                "summary": "接口分页列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "HTTP 方法",
                        "name": "method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "路径模糊",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "所属模块",
                        "name": "module",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/system/api/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需要权限：system:api:list",
                "tags": [
                    "系统管理-接口"
                ],
                "summary": "全部接口（分组展示用）",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/system/api/sync": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需要权限：system:api:update",
                "tags": [
                    "系统管理-接口"
                ],
                "summary": "手动触发一次接口同步（常规由启动时自动完成）",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/system/api/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需要权限：system:api:update",
                "tags": [
                    "系统管理-接口"
                ],
                "summary": "编辑接口名称（其余字段由代码同步，只读）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "主键",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.ApiUpdateTitleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "system.ApiUpdateTitleRequest": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string",
                    "maxLength": 64
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer {access_token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}